+++ src/Enforce/runnables/rules/size/length_equals/index.js
@@ -1,0 +1,9 @@
+// @flow
+
+function lengthEquals(value: StringOrArray, arg1: number): boolean {
+    return value.length === arg1;
+}
+
+lengthEquals.negativeForm = 'lengthNotEquals';
+
+export default lengthEquals;