+++ src/Enforce/runnables/rules/size/shorter_than/index.js
@@ -1,0 +1,7 @@
+// @flow
+
+function shorterThan(value: StringOrArray, arg1: number): boolean {
+    return value.length < arg1;
+}
+
+export default shorterThan;