+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - [Minor] `.after()` callback that can run after a specific field finished execution.
 - [Minor] new size rules (lessThan, greaterThan, lessThanOrEquals, greaterThanOrEquals).
+- [Minor] new size rules (longerThan, shorterThan, longerThanOrEquals, shorterThanOrEquals, lengthEquals).
 
 ## [6.3.4] - 2019-05-01